backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
websockets==12.0
openai==1.63.2
pandas==2.1.3
python-dotenv==1.0.0
ollama==0.1.6
openpyxl==3.1.2 